Update
Update to f64 to increase accuracy

+++ rust/src/models/battery.rs
@@ -1,6 +1,6 @@
 pub struct Battery {
 	pub level: u8,
-	pub voltage: f32,
-	pub current: f32,
+	pub voltage: f64,
+	pub current: f64,
 	pub is_charging: bool
 }
